+++ Classes/MenuCell.m
@@ -7,8 +7,6 @@
     self = [super initWithStyle:UITableViewCellStyleDefault reuseIdentifier:reuseIdentifier];
     if (self) {
 		self.selectionStyle = UITableViewCellSelectionStyleGray;
-		self.selectedBackgroundView = [[UIView alloc] initWithFrame:self.frame];
-		self.selectedBackgroundView.backgroundColor = [UIColor colorWithRed:0.111 green:0.165 blue:0.250 alpha:1.000]; //[UIColor colorWithRed:0.178 green:0.262 blue:0.397 alpha:1.000];
 		self.imageView.contentMode = UIViewContentModeScaleAspectFit;
         self.textLabel.font = [UIFont systemFontOfSize:15];
 		self.textLabel.textColor = [UIColor whiteColor];